Dose_update and Dose_update_new are test trials where I tried to add per particle motion shifts. But i was using the wrong pixel size.
Polish_fresh is the 1st version where I can get the same particles as the original stack and the reference volumes look the same.
Features.
Added a function for applying global dose weighting on particles.
Added a function to apply global shifts.
Added a function to compute weights from bfactor.star file generated from relion.

At this stage, per frames shifts are working and global dose weighting works.

Remember. The original stack we use is already dose weighted in motion corr and is generated using corrected sums. As I am using movies which are not global dose weighted that is why I need to remember that.