set envvar TMPDIR to short path
Ref: https://www.open-mpi.org/faq/?category=osx#startup-errors-with-open-mpi-2.0.x

Without this on Travis CI:

```
[Traviss-Mac-1284.local:08278] [[16381,0],0] ORTE_ERROR_LOG: Bad parameter in file orted/pmix/pmix_server.c at line 262

[Traviss-Mac-1284.local:08278] [[16381,0],0] ORTE_ERROR_LOG: Bad parameter in file ess_hnp_module.c at line 666

--------------------------------------------------------------------------

It looks like orte_init failed for some reason; your parallel process is

likely to abort.  There are many reasons that a parallel process can

fail during orte_init; some of which are due to configuration or

environment problems.  This failure appears to be an internal failure;

here's some additional information (which may only be relevant to an

Open MPI developer):

  pmix server init failed

  --> Returned value Bad parameter (-5) instead of ORTE_SUCCESS
```